feat(Payment): 결제 요청 Mock API 구현

+++ src/main/kotlin/io/concert/booking/interfaces/dto/point/PointChargeRequest.kt
@@ -1,8 +1,10 @@
 package io.concert.booking.interfaces.dto.point
 
+import java.math.BigDecimal
+
 data class PointChargeRequest(
     val userId: Long,
-    val amount: Int,
+    val amount: BigDecimal,
 ) {
 
 }
