+++ src/main/kotlin/io/concert/booking/interfaces/dto/point/PointResponse.kt
@@ -1,4 +1,6 @@
 package io.concert.booking.interfaces.dto.point
 
-data class PointResponse(val balance: Int) {
+import java.math.BigDecimal
+
+data class PointResponse(val balance: BigDecimal) {
 }
